database refacto
