+++ oc_lettings_site/admin.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
 
-from .models import Letting
-from .models import Address
-from .models import Profile
+from lettings.models import Letting
+from lettings.models import Address
+from profiles.models import Profile
 
 
 admin.site.register(Letting)
