+++ oc_lettings_site/urls.py
@@ -3,11 +3,14 @@ from django.urls import path
 
 from . import views
 
+from lettings.views import lettings_index, index, letting
+from profiles.views import profiles_index, profile
+
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('lettings/', views.lettings_index, name='lettings_index'),
-    path('lettings/<int:letting_id>/', views.letting, name='letting'),
-    path('profiles/', views.profiles_index, name='profiles_index'),
-    path('profiles/<str:username>/', views.profile, name='profile'),
+    path('', index, name='index'),
+    path('lettings/', lettings_index, name='lettings_index'),
+    path('lettings/<int:letting_id>/', letting, name='letting'),
+    path('profiles/', profiles_index, name='profiles_index'),
+    path('profiles/<str:username>/', profile, name='profile'),
     path('admin/', admin.site.urls),
 ]
